refactor: map lpToken to poolData
