+++ contracts/utils/CommonUtils.sol
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.6.12;
+pragma experimental ABIEncoderV2;
+
+library CommonUtils {
+    function concatenate(string calldata a, string calldata b) external pure returns (string memory) {
+        return string(abi.encodePacked(a, b));
+    }
+}
